+++ package.json
@@ -71,7 +71,6 @@
     "@types/graphql": "14.5.0",
     "@types/highlight.js": "9.12.3",
     "@types/markdown-it": "0.0.9",
-    "@types/react-select": "^3.0.8",
     "ava": "2.4.0",
     "awesome-typescript-loader": "5.2.1",
     "cache-loader": "2.0.1",
@@ -179,7 +178,6 @@
     "react-redux": "5.1.2",
     "react-router": "4.3.1",
     "react-router-dom": "4.3.1",
-    "react-select": "^3.0.8",
     "react-seo-meta-tags": "1.0.0",
     "react-tooltip": "3.11.1",
     "redux": "4.0.4",
